// react-module/src/openapi/models/machine-calendar-task-basic-filtered-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface MachineCalendarTaskBasicFilteredDto
 */
export interface MachineCalendarTaskBasicFilteredDto {
    /**
     * 
     * @type {number}
     * @memberof MachineCalendarTaskBasicFilteredDto
     */
    id?: any;
    /**
     * 
     * @type {Date}
     * @memberof MachineCalendarTaskBasicFilteredDto
     */
    startDate?: any;
    /**
     * 
     * @type {Date}
     * @memberof MachineCalendarTaskBasicFilteredDto
     */
    startDatePlanned?: any;
    /**
     * 
     * @type {number}
     * @memberof MachineCalendarTaskBasicFilteredDto
     */
    duration?: any;
    /**
     * 
     * @type {OperationTypeBasicDto}
     * @memberof MachineCalendarTaskBasicFilteredDto
     */
    operationType?: any;
    /**
     * 
     * @type {MachineDto}
     * @memberof MachineCalendarTaskBasicFilteredDto
     */
    machine?: any;
    /**
     * 
     * @type {Array&lt;AtomicOperationBasicDto&gt;}
     * @memberof MachineCalendarTaskBasicFilteredDto
     */
    operationList?: any;
    /**
     * 
     * @type {string}
     * @memberof MachineCalendarTaskBasicFilteredDto
     */
    planningStatus?: MachineCalendarTaskBasicFilteredDtoPlanningStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof MachineCalendarTaskBasicFilteredDto
     */
    productionStatus?: MachineCalendarTaskBasicFilteredDtoProductionStatusEnum;
    /**
     * 
     * @type {ForSavingEstimatedTime}
     * @memberof MachineCalendarTaskBasicFilteredDto
     */
    estimatedTime?: any;
    /**
     * 
     * @type {string}
     * @memberof MachineCalendarTaskBasicFilteredDto
     */
    info?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum MachineCalendarTaskBasicFilteredDtoPlanningStatusEnum {
    INIT = 'INIT',
    PLANNED = 'PLANNED',
    CONFIRMED = 'CONFIRMED',
    FINISHED = 'FINISHED',
    DELETED = 'DELETED'
}
/**
    * @export
    * @enum {string}
    */
export enum MachineCalendarTaskBasicFilteredDtoProductionStatusEnum {
    INIT = 'INIT',
    INPROGRESS = 'IN_PROGRESS',
    POSTPONED = 'POSTPONED',
    FINISHED = 'FINISHED',
    DELETED = 'DELETED',
    FAULT = 'FAULT'
}